Modification du diagramme de classes et du modèle

+++ src/app/src/main/java/fr/iut/speedjumper/entrees/Commande.java
@@ -6,5 +6,5 @@ import fr.iut.speedjumper.entites.Entite;
  * Classe permettant d'executer une commande sur une entite
  */
 public interface Commande {
-    void execute(Entite entite, float temps);
+    void execute(Entite entite, double temps);
 }
